requirements.txt
mmh3>=4.0.0
pyyaml>=6.0
transformers>=4.40.0
ujson>=5.10.0
jsonlines>=4.0.0
pytest>=8.0.0
pytest-cov>=5.0.0
black>=24.3.0
flake8>=7.0.0
mwxml
datasets
tiktoken
torch
faiss-cpu>=1.7.4
# faiss-gpu # only works on GPU
numpy>=1.23,<2.0
rank_bm25
einops>=0.6.0
